moved classification services to a new service file
