+++ talentmap_api/fsbid/views/bureau.py
@@ -16,7 +16,7 @@ import talentmap_api.fsbid.services.available_positions as ap_services
 import talentmap_api.fsbid.services.employee as empservices
 import talentmap_api.fsbid.services.common as com_services
 import talentmap_api.fsbid.services.bid as bid_services
-import talentmap_api.fsbid.services.cdo as cdo_services
+import talentmap_api.fsbid.services.classifications as classifications_services
 
 logger = logging.getLogger(__name__)
 
@@ -116,7 +116,7 @@ class FSBidBureauPositionBidsView(BaseView):
 
         for x in result:
             perdet_seq_num = x.get('emp_id')
-            x['classifications'] = cdo_services.get_client_classification(request.META['HTTP_JWT'], perdet_seq_num)
+            x['classifications'] = classifications_services.get_client_classification(request.META['HTTP_JWT'], perdet_seq_num)
             x['has_competing_rank'] = com_services.has_competing_rank(self,  perdet_seq_num, pk)
 
         mappedResult = bid_services.map_bids_handshake_status_by_cp_id(result, pk)
